+++ package.json
@@ -56,7 +56,7 @@
     "typescript": "^5.1.6",
     "vite": "^5.1.0",
     "vite-tsconfig-paths": "^4.2.1",
-    "vitest": "^2.1.3"
+    "vitest": "^2.1.5"
   },
   "engines": {
     "node": ">=20.0.0"
