+++ vitest.config.ts
@@ -1,6 +1,12 @@
 import { defineConfig } from "vitest/config"
+import path from "path"
 
 export default defineConfig({
+  resolve: {
+    alias: {
+      "~": path.resolve(__dirname, "app"),
+    },
+  },
   test: {
     environment: "jsdom",
     globals: true,
